# docs/10-client-api/Client-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EDUagent Client-API",
    "version": "5.0.0",
    "description": "前端 (Vue3) ↔ Backend (FastAPI, Port 8001) 接口规范。\n统一返回格式: {code, message, data}。\n流式接口使用 SSE (text/event-stream)。\n前端发起刷新/生成接口后收到 task_id，并通过 /tasks/{task_id} 轮询；Backend 负责创建任务、记录归属用户和任务类型；Agent Service 长任务完成后由 Backend 接收 webhook 回调，前端不直接等待 webhook。"
  },
  "servers": [
    {
      "url": "http://localhost:8001/api/v1",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/tutoring/chat": {
      "post": {
        "tags": [
          "Tutoring"
        ],
        "summary": "智能辅导对话",
        "description": "SSE 流式接口。scope=course 为课程内智能辅导，必须传 course_id 并使用课程知识库 RAG；scope=global 为全局简单 AI Bot，不读取课程知识库。Backend 全量保存 user/assistant 消息并记录 meta，Agent 每次回复前检索用户长期记忆和相关历史事实。教师不能查看学生智能辅导对话。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 流",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "EDU SSE v2 事件：workflow_started / text_delta / tool_started / tool_completed / tool_failed / plan_updated / source_refs / artifact_created / critic_completed / content_safety_reviewed / debug_log / workflow_completed / workflow_failed。tool_started 携带 tool_title/tool_category/read_only；工具结果 outcome 为 success/neutral/warning/failure；source_refs 统一使用 payload.sources[]。"
                }
              }
            }
          }
        }
      }
    },
    "/tutoring/conversations": {
      "get": {
        "tags": [
          "Tutoring"
        ],
        "summary": "对话历史列表",
        "description": "仅返回当前登录用户自己的对话；教师不可通过该接口查看学生对话。",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "course / global"
          },
          {
            "name": "course_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "课程 ID；筛选课程内对话"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "conversations": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ConversationBrief"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tutoring/conversations/{id}": {
      "get": {
        "tags": [
          "Tutoring"
        ],
        "summary": "对话历史详情",
        "description": "仅当前登录用户本人可查看该对话详情；教师不可查看学生对话。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "对话 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConversationDetail"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "用户消息"
          },
          "scope": {
            "type": "string",
            "description": "course / global，默认 course"
          },
          "course_id": {
            "type": "string",
            "nullable": true,
            "description": "课程上下文 ID；scope=course 时必填，scope=global 时为空"
          },
          "conversation_id": {
            "type": "string",
            "nullable": true,
            "description": "继续已有对话时传入；为空则创建新对话"
          }
        }
      },
...
      "ConversationBrief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "对话 ID"
          },
          "scope": {
            "type": "string",
            "description": "course / global"
          },
          "course_id": {
            "type": "string",
            "nullable": true,
            "description": "课程 ID，全局对话为 null"
          },
          "title": {
            "type": "string",
            "description": "对话标题"
          },
          "last_message": {
            "type": "string",
            "description": "最后一条消息摘要"
          },
          "message_count": {
            "type": "integer",
            "description": "消息数"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "最后更新时间"
          }
        }
      },
      "ConversationDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "对话 ID"
          },
          "scope": {
            "type": "string",
            "description": "course / global"
          },
          "course_id": {
            "type": "string",
            "nullable": true,
            "description": "课程 ID，全局对话为 null"
          },
          "title": {
            "type": "string",
            "description": "对话标题"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "description": "user / assistant"
                },
                "content": {
                  "type": "string",
                  "description": "消息内容"
                },
                "diagrams": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "内嵌图解"
                },
                "knowledge_points": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "引用的知识点"
                },
                "meta": {
                  "type": "object",
                  "description": "消息元信息，如模型、token、检索命中、安全标记"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time",
                  "description": "消息时间"
                }
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}